Update popup to get rid of messy process identifiers
`children[0].process` would show process identifiers like `s001` instead of a process name on an idling terminal. Changed close confirm message so that this doesn't happen.

I like the idea of showing the process when confirming, but I would change it to this until there's a way to reliably produce a "pretty" process name.