feat(tabby-ssh): use string token for providing SSHProfileImporter
Until now we were using a Type Token (class SSHProfileImporter) to inject importers into the SSHProfilesService. However, this made external plugins unable of providing their own importers, since the reference to SSHProfileImporter from the built-in plugins was different to the one from any external plugin.

This probably has to do with webpack bundling different "instances" of the class SSHProfileImporter between built-in and external plugins. The implementation was the same (since the class was being imported from the tabby-ssh package), but the memory reference was different, so the dependency injector wouldn't match the external ones.

By using a string as DI Token, we ensure plugins can use a token that the injector can match, since strings are compared by value instead of reference.